backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
httpx==0.27.0
beautifulsoup4==4.12.3
apscheduler==3.10.4
python-dateutil==2.9.0
passlib[bcrypt]==1.7.4
pyjwt==2.11.0
psycopg2-binary==2.9.9
